+++ docs/source/changelog.rst
@@ -21,6 +21,7 @@ New features
 - Add :class:`validate.Range` validator for validating integer ranges.
 - Add :class:`validate.Length` validator for validating length of sized types.
 - Add :class:`validate.Exclude` validator for disallowing/excluding specific values.
+- Add :class:`validate.Or` validator for OR'ing the result of multiple validators.
 - Add :class:`ErrorContext.metadata` for storing extra error information.
 - Add :attr:`GlobalConfig.warn_unsupported_types` to control behaviour on using unsupported types.
 - Add ``init_kwargs`` parameter in :class:`fields.Object` to support passing initialization arguments.